chore(journal): use sbe to serialize journal records

+++ journal/src/main/java/io/zeebe/journal/JournalRecord.java
@@ -38,7 +38,7 @@ public interface JournalRecord {
    *
    * @return checksum
    */
-  int checksum();
+  long checksum();
 
   /**
    * Application provided data of the record
